+++ lib/karafka/routing/builder.rb
@@ -27,7 +27,7 @@ module Karafka
         each do |consumer_group|
           hashed_group = consumer_group.to_h
           validation_result = Karafka::Schemas::ConsumerGroup.call(hashed_group)
-          return if validation_result.success?
+          next if validation_result.success?
           raise Errors::InvalidConfiguration, validation_result.errors
         end
       end