+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.3-wip
 - #300 - Store value in a value key and replace its content with parsed version - without root merge
+- #331 - Disallow building groups without topics
 - Instrumentation unification. Better and more consistent naming
 - Procline instrumentation for a nicer process name